[Feat] #83 - Move base model from unoffical SDK.
