+++ chzzkpy/base_model.py
@@ -23,16 +23,32 @@ SOFTWARE.
 
 from __future__ import annotations
 
-from pydantic import PrivateAttr
 from typing import Generic, Optional, TypeVar, TYPE_CHECKING
-
-from ..base_model import ChzzkModel
-from ..base_model import Content
+from pydantic import BaseModel, ConfigDict, Extra, PrivateAttr
+from pydantic.alias_generators import to_camel
 
 if TYPE_CHECKING:
     from typing import Any, Callable, Coroutine, Optional
 
 T = TypeVar("T")
+
+
+class ChzzkModel(BaseModel):
+    model_config = ConfigDict(
+        alias_generator=to_camel, frozen=True, extra=Extra.allow  # prevent exception.
+    )
+
+    @staticmethod
+    def special_date_parsing_validator(value: T) -> T:
+        if not isinstance(value, str):
+            return value
+        return value.replace("+09", "")
+
+
+class Content(ChzzkModel, Generic[T]):
+    code: int
+    message: Optional[str]
+    content: Optional[T]
 
 
 class SearchResult(ChzzkModel, Generic[T]):
